requirements.txt
# requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

sqlalchemy==2.0.34
pydantic==2.9.2

python-dotenv==1.0.1

openai==1.54.4

python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9